requirements.txt
Django==1.8.3
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.11.0
djangorestframework==3.2.3
djangorestframework-filters==0.4.0
gunicorn==19.3.0
pystache==0.5.4
psycopg2==2.5.1
static==1.1.1
static3==0.6.1
wsgiref==0.1.2
requests==2.7.0